BugMind_AI_Engine/requirements.txt
google-genai
requests
flask
flask_cors
Pillow